# Conflicts:
#	app/Http/Controllers/Api/AuthController.php

+++ database/migrations/2018_10_06_110537_create_donation_requests_table.php
@@ -20,8 +20,8 @@ class CreateDonationRequestsTable extends Migration {
 			$table->integer('city_id');
 			$table->string('phone');
 			$table->text('notes')->nullable();
-			$table->double('latitude', 10,10)->nullable();
-			$table->double('longitude', 10,10)->nullable();
+			$table->decimal('latitude', 10,10)->nullable();
+			$table->decimal('longitude', 10,10)->nullable();
 		});
 	}
